functioning prototype state, commiting to have a working baseline

+++ hashclient.h
@@ -11,14 +11,13 @@ Refrences
 https://en.wikibooks.org/wiki/C_Programming/Networking_in_UNIX
 
 */
-#DEFINE MAXIPLEN 16;
 
 union initialPayload
 {
     struct
     {
         int hash;
-        int sz;
+        long sz;
     };
     char data[sizeof(int)*2];
 };
